project 3
Movie Recommendation System is a machine learning project that suggests similar movies based on a user's input. It uses the movie's description to understand the storyline and recommends movies with similar plots. This project uses TF-IDF and cosine similarity to compare movie descriptions and is built using Python, pandas, and scikit-learn.